working login page with jwt and db auth
