+++ client/src/components/Signup/signup.js
@@ -29,7 +29,6 @@ const handleSubmit = async (e) =>{
             && err.response.status <=500){
                 setError(err.response.data.message)
             }
-
     }
 }
     return(
