+++ client/src/components/overview/Overview.js
@@ -9,7 +9,7 @@ const Overview = () => {
 	return (
 		<div className={styles.main_container}>
 			<nav className={styles.navbar}>
-				<h1>fakebook</h1>
+				<h1>Home Page</h1>
 				<button className={styles.white_btn} onClick={handleLogout}>
 					Logout
 				</button>
